Adding events and enhancing commands

+++ src/commands/embed.ts
@@ -2,6 +2,7 @@ import { Command } from "../structures/Command";
 import { VorteClient } from "../structures/VorteClient";
 import { Message } from "discord.js";
 import VorteEmbed from "../structures/VorteEmbed";
+import { checkPermissions } from "../util";
 
 export class Cmd extends Command {
   constructor(bot: VorteClient) {
@@ -12,6 +13,7 @@ export class Cmd extends Command {
     })
   }
   run(message: Message, args: string[]) {
+    if (!checkPermissions(message.member!, "ADMINISTRATOR")) return message.channel.send("You dont have permissions.");
     const emb = args.join(" ").split(" | ");
     if (!message.deletable) return message.channel.send("Dont have permission to delete the message");
     message.delete();
